T344 Fix prod v1 Alert class update form changes Fix getting learning sheet of user Fix some php version reference Fix some name length Fix class creating with new learning sheet
