+++ configs/entities/learning-category.yaml
@@ -4,6 +4,6 @@ fields:
     create_ip: ip=clientIP()
     create_user_id: ref=userID()
 
-    name: string(2, 50)
+    name: string(2, 100)
     key: string(2, 50)
     learning_sheet_id: ref(learning-sheet)
