+++ src/App/Controller/User/UserClassCreateController.php
@@ -5,6 +5,7 @@
 
 namespace App\Controller\User;
 
+use App\Entity\LearningSheet;
 use App\Entity\SchoolClass;
 use App\Entity\User;
 use Orpheus\Exception\UserException;
@@ -27,8 +28,11 @@ class UserClassCreateController extends AbstractUserController {
 		try {
 			if( $request->hasData('submitCreate') ) {
 				$classInput = $request->getData('class');
+				if( !empty($classInput['name']) && !empty($classInput['level']) && !empty($classInput['learning_sheet_id']) && $classInput['learning_sheet_id'] === 'new' ) {
+					$classInput['learning_sheet_id'] = LearningSheet::make($classInput['name'], $classInput['level']);
+				}
 				$classInput['teacher_id'] = User::getLoggedUser()->getPerson()->id();
-				$classId = SchoolClass::create($classInput, ['name', 'year', 'level', 'openDate', 'teacher_id']);
+				$classId = SchoolClass::create($classInput, ['name', 'year', 'level', 'openDate', 'teacher_id', 'learning_sheet_id']);
 				
 				return new RedirectHTTPResponse(u('user_class_edit', ['classId' => $classId]));
 			}
