+++ src/App/Entity/Person.php
@@ -7,7 +7,6 @@ namespace App\Entity;
 
 use DateTime;
 use Orpheus\EntityDescriptor\PermanentEntity;
-use Orpheus\SQLRequest\SQLSelectRequest;
 
 /**
  * Class Person
@@ -46,20 +45,6 @@ class Person extends PermanentEntity {
 		}
 		
 		return $pupilSkills;
-	}
-	
-	/**
-	 * @param string|string[],null $role
-	 * @return SQLSelectRequest
-	 */
-	public function queryLearningSheets($role = null): SQLSelectRequest {
-		$query = LearningSheet::get()
-			->join(LearningSheetUser::class, $userAlias, null, 'learning_sheet_id', true);
-		if( $role ) {
-			$query->where($userAlias . '.role', $role);
-		}
-		
-		return $query;
 	}
 	
 	public function getLabel() {
